+++ subsys/bluetooth/controller/ll_sw/nordic/hal/nrf5/radio/radio.c
@@ -22,6 +22,7 @@
 #include "ll_sw/pdu.h"
 
 #include "radio_internal.h"
+#include "radio_df.h"
 
 #if defined(CONFIG_BT_CTLR_GPIO_PA_PIN)
 #if ((CONFIG_BT_CTLR_GPIO_PA_PIN) > 31)
@@ -1507,4 +1508,39 @@ void radio_ar_resolve(uint8_t *addr)
 	NRF_AAR->ENABLE = (AAR_ENABLE_ENABLE_Disabled << AAR_ENABLE_ENABLE_Pos) &
 			  AAR_ENABLE_ENABLE_Msk;
 
+}
+
+/* @brief Function configures CTE inline register to start sampling of CTE
+ *        according to information parsed from CTEInfo field of received PDU.
+ *
+ * @param[in] cte_info_in_s1    Informs where to expect CTEInfo field in PDU:
+ *                              in S1 for data pdu, not in S1 for adv. PDU
+ */
+void radio_df_cte_inline_set_enabled(bool cte_info_in_s1)
+{
+#if defined(HAS_CTEINLINE_SUPPORT)
+	const nrf_radio_cteinline_conf_t inline_conf = {
+		.enable = true,
+		/* Indicates whether CTEInfo is in S1 byte or not. */
+		.info_in_s1 = cte_info_in_s1,
+	/* Enable or disable switching and sampling when CRC is not OK. */
+#if defined(CONFIG_BT_CTLR_DF_SAMPLE_CTE_FOR_PDU_WITH_BAD_CRC)
+		.err_handling = true,
+#else
+		.err_handling = false,
+#endif /* CONFIG_BT_CTLR_DF_SAMPLE_CTE_FOR_PDU_WITH_BAD_CRC */
+		/* Maximum range of CTE time. 20 * 8us according to BT spec.*/
+		.time_range = NRF_RADIO_CTEINLINE_TIME_RANGE_20,
+		/* Spacing between samples for 1us AoD or AoA is set to 2us. */
+		.rx1us = NRF_RADIO_CTEINLINE_RX_MODE_2US,
+		/* Spacing between samples for 2us AoD or AoA is set to 4us. */
+		.rx2us = NRF_RADIO_CTEINLINE_RX_MODE_4US,
+		/**< S0 bit pattern to match all types of adv. PDUs */
+		.s0_pattern = 0x0,
+		/**< S0 bit mask set to don't match any bit in SO octet */
+		.s0_mask = 0x0
+	};
+
+	nrf_radio_cteinline_configure(NRF_RADIO, &inline_conf);
+#endif /* HAS_CTEINLINE_SUPPORT */
 }
