+++ subsys/bluetooth/controller/ll_sw/nordic/hal/nrf5/radio/radio.h
@@ -111,3 +111,8 @@ uint32_t radio_ar_match_get(void);
 void radio_ar_status_reset(void);
 uint32_t radio_ar_has_match(void);
 void radio_ar_resolve(uint8_t *addr);
+
+/* Enables CTE inline configuration to automatically setup sampling and
+ * switching according to CTEInfo in received PDU.
+ */
+void radio_df_cte_inline_set_enabled(bool cte_info_in_s1);
